+++ lib/iron_admin/resource.rb
@@ -42,6 +42,7 @@ module IronAdmin
   class Resource
     include Concerns::Nestable
     include Concerns::SoftDeletable
+    include Concerns::LiveUpdatable
 
     class_attribute :field_overrides, default: {}
     class_attribute :_searchable_columns, default: nil